Cleanup YOLOv5GraphSurgeon
